+++ HISTORY.md
@@ -1,3 +1,10 @@
+1.1.0 / 2017-07-26
+==================
+
+* Rewrote most of the application to use promises
+* Use default postgres database
+* Use username from configuration instead of falling back to process.env.USER
+
 1.0.1 / 2014-11-26
 ==================